+++ MingKeMing/src/main/java/chat/dim/core/IDFactory.java
@@ -43,7 +43,7 @@ public final class IDFactory implements ID.Factory {
     private final Map<String, ID> identifiers = new HashMap<>();
 
     @Override
-    public ID generateID(Meta meta, byte type, String terminal) {
+    public ID generateID(Meta meta, int type, String terminal) {
         Address address = Address.generate(meta, type);
         assert address != null : "failed to generate ID with meta: " + meta.toMap();
         return ID.create(meta.getSeed(), address, terminal);
